+++ pushnotifications/src/main/java/com/pusher/pushnotifications/reporting/FCMMessageReceiver.kt
@@ -8,8 +8,10 @@ import com.google.gson.Gson
 import com.google.gson.JsonSyntaxException
 import com.pusher.pushnotifications.featureflags.FeatureFlag
 import com.pusher.pushnotifications.featureflags.FeatureFlagManager
+import com.pusher.pushnotifications.internal.AppActivityLifecycleCallbacks
 import com.pusher.pushnotifications.internal.DeviceStateStore
 import com.pusher.pushnotifications.logging.Logger
+import com.pusher.pushnotifications.reporting.api.DeliveryEvent
 import com.pusher.pushnotifications.reporting.api.ReportEvent
 import com.pusher.pushnotifications.reporting.api.ReportEventType
 
@@ -39,11 +41,13 @@ class FCMMessageReceiver : WakefulBroadcastReceiver() {
           return
         }
 
-        val reportEvent = ReportEvent(
-          eventType = ReportEventType.Delivery,
+        val reportEvent = DeliveryEvent(
           publishId = pusherData.publishId,
           deviceId =   deviceId,
-          timestampSecs = Math.round(System.currentTimeMillis() / 1000.0)
+          timestampSecs = Math.round(System.currentTimeMillis() / 1000.0),
+          appInBackground = AppActivityLifecycleCallbacks.appInBackground(),
+          hasDisplayableContent = pusherData.hasDisplayableContent,
+          hasData = pusherData.hasData
         )
 
         val dispatcher = FirebaseJobDispatcher(GooglePlayDriver(context))