+++ pushnotifications/src/main/java/com/pusher/pushnotifications/reporting/ReportingJobService.kt
@@ -3,18 +3,25 @@ package com.pusher.pushnotifications.reporting
 import android.os.Bundle
 import com.firebase.jobdispatcher.JobParameters
 import com.firebase.jobdispatcher.JobService
+import com.google.gson.annotations.SerializedName
 import com.pusher.pushnotifications.logging.Logger
 import com.pusher.pushnotifications.PushNotificationsInstance
 import com.pusher.pushnotifications.api.OperationCallback
-import com.pusher.pushnotifications.reporting.api.ReportEvent
-import com.pusher.pushnotifications.reporting.api.ReportEventType
-import com.pusher.pushnotifications.reporting.api.ReportingAPI
-import com.pusher.pushnotifications.reporting.api.UnrecoverableRuntimeException
+import com.pusher.pushnotifications.reporting.api.*
 
 data class PusherMetadata(
-        val publishId: String,
-        val clickAction: String?
-)
+  val publishId: String,
+  val clickAction: String?,
+  @SerializedName("hasDisplayableContent") private val _hasDisplayableContent: Boolean?,
+  @SerializedName("hasData") private val _hasData: Boolean?
+) {
+  val hasDisplayableContent: Boolean
+    get() = _hasDisplayableContent ?: false
+
+  val hasData: Boolean
+    get() = _hasData ?: false
+
+}
 
 class ReportingJobService: JobService() {
   companion object {
@@ -22,24 +29,55 @@ class ReportingJobService: JobService() {
     private const val BUNDLE_DEVICE_ID_KEY = "DeviceId"
     private const val BUNDLE_PUBLISH_ID_KEY = "PublishId"
     private const val BUNDLE_TIMESTAMP_KEY = "Timestamp"
+    private const val BUNDLE_APP_IN_BACKGROUND_KEY = "AppInBackground"
+    private const val BUNDLE_HAS_DISPLAYABLE_CONTENT_KEY = "HasDisplayableContent"
+    private const val BUNDLE_HAS_DATA_KEY = "HasData"
 
     fun toBundle(reportEvent: ReportEvent): Bundle {
       val b = Bundle()
-      b.putString(BUNDLE_EVENT_TYPE_KEY, reportEvent.eventType.toString())
-      b.putString(BUNDLE_DEVICE_ID_KEY, reportEvent.deviceId)
-      b.putString(BUNDLE_PUBLISH_ID_KEY, reportEvent.publishId)
-      b.putLong(BUNDLE_TIMESTAMP_KEY, reportEvent.timestampSecs)
+      when (reportEvent) {
+        is DeliveryEvent -> {
+          b.putString(BUNDLE_EVENT_TYPE_KEY, reportEvent.event.toString())
+          b.putString(BUNDLE_DEVICE_ID_KEY, reportEvent.deviceId)
+          b.putString(BUNDLE_PUBLISH_ID_KEY, reportEvent.publishId)
+          b.putLong(BUNDLE_TIMESTAMP_KEY, reportEvent.timestampSecs)
+          b.putBoolean(BUNDLE_APP_IN_BACKGROUND_KEY, reportEvent.appInBackground!!)
+          b.putBoolean(BUNDLE_HAS_DISPLAYABLE_CONTENT_KEY, reportEvent.hasDisplayableContent!!)
+          b.putBoolean(BUNDLE_HAS_DATA_KEY, reportEvent.hasData!!)
+        }
+
+        is OpenEvent -> {
+          b.putString(BUNDLE_EVENT_TYPE_KEY, reportEvent.event.toString())
+          b.putString(BUNDLE_DEVICE_ID_KEY, reportEvent.deviceId)
+          b.putString(BUNDLE_PUBLISH_ID_KEY, reportEvent.publishId)
+          b.putLong(BUNDLE_TIMESTAMP_KEY, reportEvent.timestampSecs)
+        }
+      }
 
       return b
     }
 
     fun fromBundle(bundle: Bundle): ReportEvent {
-      return ReportEvent(
-        eventType =  ReportEventType.valueOf(bundle.getString(BUNDLE_EVENT_TYPE_KEY)),
-        deviceId = bundle.getString(BUNDLE_DEVICE_ID_KEY),
-        publishId = bundle.getString(BUNDLE_PUBLISH_ID_KEY),
-        timestampSecs = bundle.getLong(BUNDLE_TIMESTAMP_KEY)
-      )
+      val event = ReportEventType.valueOf(bundle.getString(BUNDLE_EVENT_TYPE_KEY))
+      when (event) {
+        ReportEventType.Delivery -> {
+          return DeliveryEvent(
+            deviceId = bundle.getString(BUNDLE_DEVICE_ID_KEY),
+            publishId = bundle.getString(BUNDLE_PUBLISH_ID_KEY),
+            timestampSecs = bundle.getLong(BUNDLE_TIMESTAMP_KEY),
+            appInBackground = bundle.getBoolean(BUNDLE_APP_IN_BACKGROUND_KEY),
+            hasDisplayableContent = bundle.getBoolean(BUNDLE_HAS_DISPLAYABLE_CONTENT_KEY),
+            hasData = bundle.getBoolean(BUNDLE_HAS_DATA_KEY)
+          )
+        }
+        ReportEventType.Open -> {
+          return OpenEvent(
+            deviceId = bundle.getString(BUNDLE_DEVICE_ID_KEY),
+            publishId = bundle.getString(BUNDLE_PUBLISH_ID_KEY),
+            timestampSecs = bundle.getLong(BUNDLE_TIMESTAMP_KEY)
+          )
+        }
+      }
     }
   }