+++ pushnotifications/src/main/java/com/pusher/pushnotifications/reporting/api/ReportingService.kt
@@ -7,24 +7,36 @@ interface ReportingService {
   @POST("instances/{instanceId}/events")
   fun submit(
     @Path("instanceId") instanceId: String,
-    @Body reportingRequest: ReportingRequest
+    @Body reportingRequest: ReportEvent
   ): Call<Void>
 }
-
-data class ReportingRequest(
-  val eventType: String,
-  val publishId: String,
-  val deviceId: String,
-  val timestampSecs: Long
-)
 
 enum class ReportEventType {
   Delivery, Open,
 }
 
-data class ReportEvent(
-  val eventType: ReportEventType,
+sealed class ReportEvent(
+  val event: ReportEventType,
   val deviceId: String,
   val publishId: String,
-  val timestampSecs: Long
+  val timestampSecs: Long,
+  val appInBackground: Boolean? = null,
+  val hasDisplayableContent: Boolean? = null,
+  val hasData: Boolean? = null
 )
+
+class OpenEvent (
+  deviceId: String,
+  publishId: String,
+  timestampSecs: Long
+): ReportEvent(ReportEventType.Open, deviceId, publishId, timestampSecs)
+
+class DeliveryEvent (
+  deviceId: String,
+  publishId: String,
+  timestampSecs: Long,
+  appInBackground: Boolean,
+  hasDisplayableContent: Boolean,
+  hasData: Boolean
+): ReportEvent(ReportEventType.Delivery, deviceId, publishId, timestampSecs, appInBackground, hasDisplayableContent, hasData)
+